requirements.txt
streamlit
opencv-python
requests
python-dotenv
huggingface_hub
torch
transformers
accelerate
bitsandbytes
Pillow
